+++ webapp/src/components/ConfigurationChannelRoleSpecificPermissions.js
