+++ webapp/src/components/ConfigurationPageChannelPermissions.js
