+++ webapp/src/pages/LoginPage.js
@@ -80,7 +80,6 @@ const LoginPage = () => {
             autoComplete="username"
             autoFocus
             value={username}
-            // sx={{ backgroundColor: "primary.light", color: "text.primary" }}
             onChange={(e) => setUsername(e.target.value)}
           />
           <TextField
@@ -93,7 +92,6 @@ const LoginPage = () => {
             id="password"
             autoComplete="current-password"
             value={password}
-            // sx={{ backgroundColor: "primary.light", color: "text.primary" }}
             onChange={(e) => setPassword(e.target.value)}
           />
           <Button
@@ -103,8 +101,6 @@ const LoginPage = () => {
             sx={{
               mt: 3,
               mb: 2,
-              // backgroundColor: "primary.dark",
-              // color: "text.primary",
             }}
           >
             Sign In
@@ -114,9 +110,7 @@ const LoginPage = () => {
             variant="contained"
             sx={{
               mt: 3,
-              mb: 2,
-              // backgroundColor: "primary.dark",
-              // color: "text.primary",
+              mb: 2
             }}
             onClick={handleRegisterSubmit}
             type="button"
